Add files via upload
The full t-SQL query I'll be using to generate the data set that I'll train the TP Enrollment Forecasting model on. This version is far superior to the last each case is a unique month, year, territory partner combination as that is the level of thing that I'm trying to predict on. My previous version of this query, while much simpler to write, should only be utilized for the naive linear projection version of the forecast I built in Excel. It will not allow me to correctly factor in seasonality or my TPP and TreX flags. I may place this in an sse view in the SQL Server later on to avoid the temp table error with pyodbc.