Add files via upload
Using negative binomial regression rather than Poisson, because the variance is about 100x the mean. I also started tooling around with a bunch of Seaborn pairplots in this workbook, which I had seen a similar thing in R before, but this was my first time really using them in Python. Most of the plots are meaningless, but some interesting relationships do appear.