Add files via upload
This update to the model includes a negative binomial regression as well as a bunch of interesting seaborn pairplots. The pairplots are mostly meaningless, but some of them are interesting to look at. A couple outliers really stand out here. I didn't use the initially proposed Poisson regression because the variance was so much higher than the mean.